% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{duckknit_kill_session}
\alias{duckknit_kill_session}
\title{Kill a DuckDB CLI session}
\usage{
duckknit_kill_session(name)
}
\arguments{
\item{name}{Character. Name of the session to kill.}
}
\description{
Kill a DuckDB CLI session
}
